adding test classes and modefied error messages to be clear

+++ src/main/java/com/mytask/transfermoney/AccountService.java
@@ -4,6 +4,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class AccountService {
@@ -20,36 +21,64 @@ public class AccountService {
         return accounts;
     }
 
-    public Account addNewAccount(Account account) {
-        return accountRepository.save(account);
+    public String addNewAccount(Account account) {
+        if (!accountRepository.existsById(account.getClientNumber())) {
+            accountRepository.save(account);
+            return "Done , code = 200 \nNew Account Added Successfully\n" + account.toString();
+        }
+        throw AccountException.alreadyExist(account.getClientNumber());
     }
 
-    public Account getAccount(Long id) {
-        return accountRepository.findById(id)
-                .orElseThrow(() -> new AccountNotFoundException(id));
+    public String getAccount(Long id) {
+        if (accountRepository.existsById(id)) {
+            accountRepository.findById(id);
+            return "Done , code = 200 \nThe Account Information is \n" +
+                    accountRepository.findById(id).toString();
+        }
+        throw AccountException.notFound(id);
     }
 
-    public Account updateAccount(Account account) {
-        return accountRepository.save(account);
+    public String updateAccount(Account account) {
+        if (accountRepository.existsById(account.getClientNumber())) {
+            accountRepository.save(account);
+            return "Done , code = 200\nThe New Account Information is\n" +
+                    account.toString();
+        }
+        throw AccountException.notFound(account.getClientNumber());
     }
 
-    public void removeAccount(Long id) {
-        accountRepository.deleteById(id);
+    public String removeAccount(Long id) {
+        if (accountRepository.existsById(id)) {
+            accountRepository.deleteById(id);
+            return "Done , code = 200\nThe Account { " + id + " } was deleted";
+        }
+        throw AccountException.notFound(id);
     }
 
     public String transferMoney(Long sender, Long receiver, Double amount, String description) {
-        Account senderAcc = getAccount(sender);
-        Account receiverAcc = getAccount(receiver);
+        if (accountRepository.existsById(sender) && accountRepository.existsById(receiver)) {
 
-        Double senderAmount = senderAcc.getClientBalance();
-        Double receiverAmount = receiverAcc.getClientBalance();
+            Optional<Account> senderAcc = accountRepository.findById(sender);
+            Optional<Account> receiverAcc = accountRepository.findById(receiver);
 
-        senderAcc.setClientBalance(senderAmount - amount);
-        receiverAcc.setClientBalance(receiverAmount + amount);
+            Double senderAmount = senderAcc.get().getClientBalance();
+            Double receiverAmount = receiverAcc.get().getClientBalance();
+            try {
+                senderAcc.get().setClientBalance(senderAmount - amount);
+                receiverAcc.get().setClientBalance(receiverAmount + amount);
 
-        updateAccount(senderAcc);
-        updateAccount(receiverAcc);
+                updateAccount(senderAcc.get());
+                updateAccount(receiverAcc.get());
 
-        return "DONE\n" + description;
+            } catch (RuntimeException e) {
+                throw new RuntimeException(e);
+            }
+            return "DONE\n" + description;
+        } else {
+            Long wrongID;
+            if (accountRepository.existsById(sender)) wrongID = receiver;
+            else wrongID = sender;
+            throw AccountException.notFound(wrongID);
+        }
     }
 }
